refactor(earn): share yield approval action logic
