+++ suite-common/wallet-core/src/index.ts
@@ -61,6 +61,7 @@ export * from './stake/stakeSelectors';
 export * from './stake/stakeThunks';
 export * from './stablecoin-yield/stablecoinYieldReducer';
 export * from './stablecoin-yield/stablecoinYieldSelectors';
+export * from './stablecoin-yield/stablecoinYieldApprovalActionUtils';
 export * from './stablecoin-yield/stablecoinYieldApprovalThunks';
 export * from './stablecoin-yield/stablecoinYieldDepositThunks';
 export * from './stablecoin-yield/stablecoinYieldTypes';
